third_party/WebKit/LayoutTests/platform/mac-mountainlion/http/tests/security/suborigins/suborigin-invalid-names-expected.txt

ALERT: I am a secret
CONSOLE ERROR: Error while parsing the 'suborigin' Content Security Policy directive: Invalid character '\' in suborigin.
ALERT: I am a secret
CONSOLE ERROR: Error while parsing the 'suborigin' Content Security Policy directive: Invalid character '\' in suborigin.
ALERT: I am a secret
CONSOLE ERROR: Error while parsing the 'suborigin' Content Security Policy directive: Invalid character '\' in suborigin.
ALERT: I am a secret
CONSOLE ERROR: Error while parsing the 'suborigin' Content Security Policy directive: Invalid character '\' in suborigin.
ALERT: I am a secret
CONSOLE ERROR: Error while parsing the 'suborigin' Content Security Policy directive: Invalid character '@' in suborigin.
ALERT: I am a secret
CONSOLE ERROR: Error while parsing the 'suborigin' Content Security Policy directive: Whitespace is not allowed in suborigin names.
ALERT: I am a secret
